pyproject.toml
[project]
name = "rasp-aqua"
version = "0.1.0"
description = "水槽を Raspberry Pi でモニターしつつ管理するスクリプトです．"
authors = [
    { name = "KIMATA Tetsuya", email = "kimata@green-rabbit.net" }
]
readme = "README.md"
requires-python = ">= 3.10"

dependencies = [
    "coloredlogs>=15.0.1",
    "docopt-ng>=0.9.0",
    "pytz>=2024.1",
    "pyyaml>=6.0.1",
    "schedule>=1.2.2",
    "rpi-lgpio>=0.6",
    "influxdb-client[ciso]>=1.44.0",
    "my-lib @ git+https://github.com/kimata/my-py-lib@c93ad83b845ec083bb021b2866b7a7063bcc733c",
    "opencv-contrib-python-headless>=4.10.0.84",
    "lxml>=5.3.0",
    "matplotlib>=3.9.2",
    "pandas>=2.2.2",
    "selenium>=4.23.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rasp_aqua"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.rye]
managed = true
dev-dependencies = [
    "flaky>=3.8.1",
    "playwright>=1.45.1",
    "pytest-cov>=5.0.0",
    "pytest-html>=4.1.1",
    "pytest-mock>=3.14.0",
    "pytest-playwright>=0.5.1",
    "pytest-xdist>=3.6.1",
    "time-machine>=2.15.0",
]

[tool.pytest.ini_options]
minversion = "6.0"

# addopts = "--verbose --log-cli-level=DEBUG --log-file-level=DEBUG --log-format=\"%(asctime)s %(levelname)s %(message)s\" --log-format=\"%(asctime)s %(levelname)s [%(filename)s:%(lineno)s %(funcName)s] %(message)s\" --capture=sys  --html=tests/evidence/index.htm --self-contained-html --cov=src --cov-report=html"

addopts = "--verbose --log-file-level=DEBUG --log-format=\"%(asctime)s %(levelname)s %(message)s\" --log-format=\"%(asctime)s %(levelname)s [%(filename)s:%(lineno)s %(funcName)s] %(message)s\" --capture=sys  --html=tests/evidence/index.htm --self-contained-html --cov=src --cov-report=html"

testpaths = [
    "tests",
]
filterwarnings = [
    "ignore:The hookimpl CovPlugin.pytest_configure_node uses",
    "ignore:The hookimpl CovPlugin.pytest_testnodedown uses",
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",

  # 以下，未使用関数
]

[tool.coverage.html]
directory = "tests/evidence/coverage"
